Update TransactionService.php
This changes how additionals are handled. They are all sent to Admin as a centralizing account, from where all the rest of distributions start cascading. This to connect all accounts so all the activity of the network contributes to updating balances. 

There's also some lines that needed formatting with Tab. so adjusted them, they have the same code.